Got bullets spawning, BUT.
Bullets spawn and copy the rotation at the time of spawning, but
otherwise have no other functionality at the moment.

IDEA:
Have a generic script for all bullets, but by default all values are
null/0. The values are set at the time of creation when the player
fires.

Explain: This way I can just make 1 general script for the bullets of
every gun, but simply have valves be assigned or not assigned.